pyproject.toml
[tool.poetry]
name = "tombo"
version = "0.1.0"
description = ""
authors = ["bruno santiago <bruno.queiroz@ssp.ba.gov.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.105.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pip = "^23.3.2"
install = "^1.3.5"
pydantic = "^2.5.3"


[tool.poetry.group.dev.dependencies]
ipython = "^8.18.1"
ipykernel = "^6.27.1"
isort = "^5.13.2"
black = "^23.12.0"
ruff = "^0.1.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
